+++ src/parallel_fann_cpp.cpp
@@ -9,7 +9,7 @@ using namespace std;
 namespace parallel_fann {
 // TODO rewrite all these functions in c++ using fann_cpp interface
 
-float train_epoch_batch_parallel(struct fann *ann, struct fann_train_data *data, const unsigned int threadnumb)
+double train_epoch_batch_parallel(struct fann *ann, struct fann_train_data *data, const unsigned int threadnumb)
 {
 	fann_reset_MSE(ann);
 	vector<struct fann *> ann_vect(threadnumb);
@@ -78,7 +78,7 @@ float train_epoch_batch_parallel(struct fann *ann, struct fann_train_data *data,
 }
 
 
-float train_epoch_irpropm_parallel(struct fann *ann, struct fann_train_data *data, const unsigned int threadnumb)
+double train_epoch_irpropm_parallel(struct fann *ann, struct fann_train_data *data, const unsigned int threadnumb)
 {
 
 	if(ann->prev_train_slopes == NULL)
@@ -125,10 +125,10 @@ float train_epoch_irpropm_parallel(struct fann *ann, struct fann_train_data *dat
 
     	fann_type next_step;
 
-    	const float increase_factor = ann->rprop_increase_factor;	//1.2;
-    	const float decrease_factor = ann->rprop_decrease_factor;	//0.5;
-    	const float delta_min = ann->rprop_delta_min;	//0.0;
-    	const float delta_max = ann->rprop_delta_max;	//50.0;
+    	const double increase_factor = ann->rprop_increase_factor;	//1.2;
+    	const double decrease_factor = ann->rprop_decrease_factor;	//0.5;
+    	const double delta_min = ann->rprop_delta_min;	//0.0;
+    	const double delta_max = ann->rprop_delta_max;	//50.0;
 		const unsigned int first_weight=0;
 		const unsigned int past_end=ann->total_connections;
 
@@ -196,7 +196,7 @@ float train_epoch_irpropm_parallel(struct fann *ann, struct fann_train_data *dat
 }
 
 
-float train_epoch_quickprop_parallel(struct fann *ann, struct fann_train_data *data, const unsigned int threadnumb)
+double train_epoch_quickprop_parallel(struct fann *ann, struct fann_train_data *data, const unsigned int threadnumb)
 {
 
 	if(ann->prev_train_slopes == NULL)
@@ -244,10 +244,10 @@ float train_epoch_quickprop_parallel(struct fann *ann, struct fann_train_data *d
 
     	fann_type w=0.0, next_step;
 
-    	const float epsilon = ann->learning_rate / data->num_data;
-    	const float decay = ann->quickprop_decay;	/*-0.0001;*/
-    	const float mu = ann->quickprop_mu;	/*1.75; */
-    	const float shrink_factor = (float) (mu / (1.0 + mu));
+    	const double epsilon = ann->learning_rate / data->num_data;
+    	const double decay = ann->quickprop_decay;	/*-0.0001;*/
+    	const double mu = ann->quickprop_mu;	/*1.75; */
+    	const double shrink_factor = (double) (mu / (1.0 + mu));
 
 		omp_set_dynamic(0);
 		omp_set_num_threads(threadnumb);
@@ -330,7 +330,7 @@ float train_epoch_quickprop_parallel(struct fann *ann, struct fann_train_data *d
 }
 
 
-float train_epoch_sarprop_parallel(struct fann *ann, struct fann_train_data *data, const unsigned int threadnumb)
+double train_epoch_sarprop_parallel(struct fann *ann, struct fann_train_data *data, const unsigned int threadnumb)
 {
 	if(ann->prev_train_slopes == NULL)
 	{
@@ -379,15 +379,15 @@ float train_epoch_sarprop_parallel(struct fann *ann, struct fann_train_data *dat
     	fann_type next_step;
 
     	/* These should be set from variables */
-    	const float increase_factor = ann->rprop_increase_factor;	/*1.2; */
-    	const float decrease_factor = ann->rprop_decrease_factor;	/*0.5; */
+    	const double increase_factor = ann->rprop_increase_factor;	/*1.2; */
+    	const double decrease_factor = ann->rprop_decrease_factor;	/*0.5; */
     	/* TODO: why is delta_min 0.0 in iRprop? SARPROP uses 1x10^-6 (Braun and Riedmiller, 1993) */
-    	const float delta_min = 0.000001f;
-    	const float delta_max = ann->rprop_delta_max;	/*50.0; */
-    	const float weight_decay_shift = ann->sarprop_weight_decay_shift; /* ld 0.01 = -6.644 */
-    	const float step_error_threshold_factor = ann->sarprop_step_error_threshold_factor; /* 0.1 */
-    	const float step_error_shift = ann->sarprop_step_error_shift; /* ld 3 = 1.585 */
-    	const float T = ann->sarprop_temperature;
+    	const double delta_min = 0.000001f;
+    	const double delta_max = ann->rprop_delta_max;	/*50.0; */
+    	const double weight_decay_shift = ann->sarprop_weight_decay_shift; /* ld 0.01 = -6.644 */
+    	const double step_error_threshold_factor = ann->sarprop_step_error_threshold_factor; /* 0.1 */
+    	const double step_error_shift = ann->sarprop_step_error_shift; /* ld 3 = 1.585 */
+    	const double T = ann->sarprop_temperature;
 
 
     	//merge of MSEs
@@ -397,8 +397,8 @@ float train_epoch_sarprop_parallel(struct fann *ann, struct fann_train_data *dat
     		ann->num_MSE+=ann_vect[i]->num_MSE;
     	}
 
-    	const float MSE = fann_get_MSE(ann);
-    	const float RMSE = sqrtf(MSE);
+    	const double MSE = fann_get_MSE(ann);
+    	const double RMSE = sqrt(MSE);
 
     	/* for all weights; TODO: are biases included? */
 		omp_set_dynamic(0);
@@ -446,7 +446,7 @@ float train_epoch_sarprop_parallel(struct fann *ann, struct fann_train_data *dat
 						#define	RAND_MAX	0x7fffffff
 						#endif
 						if(prev_step < step_error_threshold_factor * MSE)
-							next_step = prev_step * decrease_factor + (float)rand() / RAND_MAX * RMSE * (fann_type)fann_exp2(-T * epoch + step_error_shift);
+							next_step = prev_step * decrease_factor + (double)rand() / RAND_MAX * RMSE * (fann_type)fann_exp2(-T * epoch + step_error_shift);
 						else
 							next_step = fann_max(prev_step * decrease_factor, delta_min);
 
@@ -485,7 +485,7 @@ float train_epoch_sarprop_parallel(struct fann *ann, struct fann_train_data *dat
 	return fann_get_MSE(ann);
 }
 
-float train_epoch_incremental_mod(struct fann *ann, struct fann_train_data *data)
+double train_epoch_incremental_mod(struct fann *ann, struct fann_train_data *data)
 {
 	unsigned int i;
 
@@ -501,7 +501,7 @@ float train_epoch_incremental_mod(struct fann *ann, struct fann_train_data *data
 
 //the following versions returns also the outputs via the predicted_outputs parameter
 
-float train_epoch_batch_parallel(struct fann *ann, struct fann_train_data *data, const unsigned int threadnumb,vector< vector<fann_type> >& predicted_outputs)
+double train_epoch_batch_parallel(struct fann *ann, struct fann_train_data *data, const unsigned int threadnumb,vector< vector<fann_type> >& predicted_outputs)
 {
 	fann_reset_MSE(ann);
 	predicted_outputs.resize(data->num_data,vector<fann_type> (data->num_output));
@@ -577,7 +577,7 @@ float train_epoch_batch_parallel(struct fann *ann, struct fann_train_data *data,
 }
 
 
-float train_epoch_irpropm_parallel(struct fann *ann, struct fann_train_data *data, const unsigned int threadnumb, vector< vector<fann_type> >& predicted_outputs)
+double train_epoch_irpropm_parallel(struct fann *ann, struct fann_train_data *data, const unsigned int threadnumb, vector< vector<fann_type> >& predicted_outputs)
 {
 
 	if(ann->prev_train_slopes == NULL)
@@ -628,10 +628,10 @@ float train_epoch_irpropm_parallel(struct fann *ann, struct fann_train_data *dat
 
     	fann_type next_step;
 
-    	const float increase_factor = ann->rprop_increase_factor;	//1.2;
-    	const float decrease_factor = ann->rprop_decrease_factor;	//0.5;
-    	const float delta_min = ann->rprop_delta_min;	//0.0;
-    	const float delta_max = ann->rprop_delta_max;	//50.0;
+    	const double increase_factor = ann->rprop_increase_factor;	//1.2;
+    	const double decrease_factor = ann->rprop_decrease_factor;	//0.5;
+    	const double delta_min = ann->rprop_delta_min;	//0.0;
+    	const double delta_max = ann->rprop_delta_max;	//50.0;
 		const unsigned int first_weight=0;
 		const unsigned int past_end=ann->total_connections;
 
@@ -699,7 +699,7 @@ float train_epoch_irpropm_parallel(struct fann *ann, struct fann_train_data *dat
 }
 
 
-float train_epoch_quickprop_parallel(struct fann *ann, struct fann_train_data *data, const unsigned int threadnumb, vector< vector<fann_type> >& predicted_outputs)
+double train_epoch_quickprop_parallel(struct fann *ann, struct fann_train_data *data, const unsigned int threadnumb, vector< vector<fann_type> >& predicted_outputs)
 {
 
 	if(ann->prev_train_slopes == NULL)
@@ -752,10 +752,10 @@ float train_epoch_quickprop_parallel(struct fann *ann, struct fann_train_data *d
 
     	fann_type w=0.0, next_step;
 
-    	const float epsilon = ann->learning_rate / data->num_data;
-    	const float decay = ann->quickprop_decay;	/*-0.0001;*/
-    	const float mu = ann->quickprop_mu;	/*1.75; */
-    	const float shrink_factor = (float) (mu / (1.0 + mu));
+    	const double epsilon = ann->learning_rate / data->num_data;
+    	const double decay = ann->quickprop_decay;	/*-0.0001;*/
+    	const double mu = ann->quickprop_mu;	/*1.75; */
+    	const double shrink_factor = (double) (mu / (1.0 + mu));
 
 		omp_set_dynamic(0);
 		omp_set_num_threads(threadnumb);
@@ -838,7 +838,7 @@ float train_epoch_quickprop_parallel(struct fann *ann, struct fann_train_data *d
 }
 
 
-float train_epoch_sarprop_parallel(struct fann *ann, struct fann_train_data *data, const unsigned int threadnumb, vector< vector<fann_type> >& predicted_outputs)
+double train_epoch_sarprop_parallel(struct fann *ann, struct fann_train_data *data, const unsigned int threadnumb, vector< vector<fann_type> >& predicted_outputs)
 {
 
 	if(ann->prev_train_slopes == NULL)
@@ -893,15 +893,15 @@ float train_epoch_sarprop_parallel(struct fann *ann, struct fann_train_data *dat
     	fann_type next_step;
 
     	/* These should be set from variables */
-    	const float increase_factor = ann->rprop_increase_factor;	/*1.2; */
-    	const float decrease_factor = ann->rprop_decrease_factor;	/*0.5; */
+    	const double increase_factor = ann->rprop_increase_factor;	/*1.2; */
+    	const double decrease_factor = ann->rprop_decrease_factor;	/*0.5; */
     	/* TODO: why is delta_min 0.0 in iRprop? SARPROP uses 1x10^-6 (Braun and Riedmiller, 1993) */
-    	const float delta_min = 0.000001f;
-    	const float delta_max = ann->rprop_delta_max;	/*50.0; */
-    	const float weight_decay_shift = ann->sarprop_weight_decay_shift; /* ld 0.01 = -6.644 */
-    	const float step_error_threshold_factor = ann->sarprop_step_error_threshold_factor; /* 0.1 */
-    	const float step_error_shift = ann->sarprop_step_error_shift; /* ld 3 = 1.585 */
-    	const float T = ann->sarprop_temperature;
+    	const double delta_min = 0.000001f;
+    	const double delta_max = ann->rprop_delta_max;	/*50.0; */
+    	const double weight_decay_shift = ann->sarprop_weight_decay_shift; /* ld 0.01 = -6.644 */
+    	const double step_error_threshold_factor = ann->sarprop_step_error_threshold_factor; /* 0.1 */
+    	const double step_error_shift = ann->sarprop_step_error_shift; /* ld 3 = 1.585 */
+    	const double T = ann->sarprop_temperature;
 
 
     	//merge of MSEs
@@ -911,8 +911,8 @@ float train_epoch_sarprop_parallel(struct fann *ann, struct fann_train_data *dat
     		ann->num_MSE+=ann_vect[i]->num_MSE;
     	}
 
-    	const float MSE = fann_get_MSE(ann);
-    	const float RMSE = (float)sqrt(MSE);
+    	const double MSE = fann_get_MSE(ann);
+    	const double RMSE = (double)sqrt(MSE);
 
     	/* for all weights; TODO: are biases included? */
 		omp_set_dynamic(0);
@@ -960,7 +960,7 @@ float train_epoch_sarprop_parallel(struct fann *ann, struct fann_train_data *dat
 						#define	RAND_MAX	0x7fffffff
 						#endif
 						if(prev_step < step_error_threshold_factor * MSE)
-							next_step = prev_step * decrease_factor + (float)rand() / RAND_MAX * RMSE * (fann_type)fann_exp2(-T * epoch + step_error_shift);
+							next_step = prev_step * decrease_factor + (double)rand() / RAND_MAX * RMSE * (fann_type)fann_exp2(-T * epoch + step_error_shift);
 						else
 							next_step = fann_max(prev_step * decrease_factor, delta_min);
 
@@ -1000,7 +1000,7 @@ float train_epoch_sarprop_parallel(struct fann *ann, struct fann_train_data *dat
 }
 
 
-float train_epoch_incremental_mod(struct fann *ann, struct fann_train_data *data, vector< vector<fann_type> >& predicted_outputs)
+double train_epoch_incremental_mod(struct fann *ann, struct fann_train_data *data, vector< vector<fann_type> >& predicted_outputs)
 {
 
 	predicted_outputs.resize(data->num_data,vector<fann_type> (data->num_output));
@@ -1024,7 +1024,7 @@ float train_epoch_incremental_mod(struct fann *ann, struct fann_train_data *data
 	return fann_get_MSE(ann);
 }
 
-float test_data_parallel(struct fann *ann, struct fann_train_data *data, const unsigned int threadnumb)
+double test_data_parallel(struct fann *ann, struct fann_train_data *data, const unsigned int threadnumb)
 {
 	if(fann_check_input_output_sizes(ann, data) == -1)
 		return 0;
@@ -1064,7 +1064,7 @@ float test_data_parallel(struct fann *ann, struct fann_train_data *data, const u
 	return fann_get_MSE(ann);
 }
 
-float test_data_parallel(struct fann *ann, struct fann_train_data *data, const unsigned int threadnumb, vector< vector<fann_type> >& predicted_outputs)
+double test_data_parallel(struct fann *ann, struct fann_train_data *data, const unsigned int threadnumb, vector< vector<fann_type> >& predicted_outputs)
 {
 	if(fann_check_input_output_sizes(ann, data) == -1)
 		return 0;
